pyproject.toml
[project]
name = "strat-builder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.15",
    "alembic>=1.16.5",
    "asyncpg>=0.30.0",
    "fastapi>=0.116.1",
    "matplotlib>=3.10.6",
    "metatrader5>=5.0.5200",
    "pandas>=2.3.2",
    "psycopg2>=2.9.10",
    "pydantic>=2.11.7",
    "pyjwt>=2.10.1",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
]


[tool.pytest.ini_options]
pythonpath=['src']
